Move controllers to the `internal` directory
According to the migration doc, controllers should now
exists in the `internal` directory

https://book.kubebuilder.io/migration/manually_migration_guide_gov3_to_gov4.html

Issue: https://github.com/db-operator/db-operator/issues/3i4
