+++ internal/controller/database_helper.go
